refactor: Phase 13 — extract str_preview + 4 turn modules to crates
- Move str_preview.rs to astra-text-utils (50 LOC)
- Move safety_middleware, view, boost_domain_hints, headless_tool_status_display to astra-turn-core
- astra-turn-core: 663 → 778 tests (+115)
- Runtime: 4,172 → 4,052 tests (transferred to crates)
- Add astra-pipeline, astra-sandbox, astra-text-utils deps to astra-turn-core
