+++ rust/crates/astra-text-utils/src/lib.rs
@@ -5,4 +5,5 @@
 
 pub mod output_style;
 pub mod semantic_dedup;
+pub mod str_preview;
 pub mod text_tokenize;
