+++ rust/crates/astra-turn-core/src/lib.rs
@@ -6,9 +6,9 @@
 
 pub mod agent_progress_ui;
 pub mod agentic_verdict_audit;
+pub mod boost_domain_hints;
 pub mod cache;
-pub mod cache_diagnostics;
-pub mod chat_history_openai;
+pub mod cache_diagnostics;pub mod chat_history_openai;
 pub mod chat_turn_api_error;
 pub mod chat_turn_explain_wire;
 pub mod confidence_contract;
@@ -22,6 +22,7 @@ pub mod file_edit_journal;
 pub mod firewall;
 pub mod followup_suggestion;
 pub mod headless_tool_assembly;
+pub mod headless_tool_status_display;
 pub mod headless_tool_stderr_lines;
 pub mod history;
 pub mod hook_plans;
@@ -33,6 +34,7 @@ pub mod prepare_turn_explain_text;
 pub mod refresh;
 pub mod response_guard;
 pub mod routing;
+pub mod safety_middleware;
 pub mod skill_instructions_merge;
 pub mod snapshot;
 pub mod sse_edge_stderr_lines;
@@ -46,4 +48,5 @@ pub mod tool_result_semantics;
 pub mod tool_result_storage;
 pub mod tool_selection;
 pub mod unconsumed;
+pub mod view;
 pub mod xml_tool_call_fallback;
